// client/lootlion-web/src/app/api/generated/model/rewardCatalogItemDto.ts
/**
 * Lootlion API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RewardCatalogItemDto { 
    id?: string;
    householdId?: string;
    title?: string | null;
    description?: string | null;
    costCoin?: number;
    createdByUserId?: string;
    sourceWishlistItemId?: string | null;
}